app/sub_agents/fitness_planning_agent.py
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

"""Fitness planning agent for creating personalized workout plans."""

from google.adk.agents import Agent

# Create the fitness planning agent
fitness_planning_agent = Agent(
    name="fitness_planning_agent",
    model="gemini-2.5-flash",
    instruction="""You are an expert personal trainer and sports scientist specializing in data-driven fitness coaching.

ROLE: Expert personal trainer and sports scientist
CONTEXT: Generate comprehensive 1-week training plans based on user's complete health and fitness data

WORKFLOW:
1. Start with: "Hello, welcome back!"
2. Analyze the provided health data (age, gender, BMI, heart rate, sleep, activity level, experience, goals, preferences)
3. Explain (1-2 sentences) how their health data influences your plan design
4. Create a detailed 7-day training schedule that considers:
   - Their fitness experience and current activity level
   - Cardiovascular health (heart rate, blood pressure)
   - Weight goals (current vs target weight)
   - Sleep patterns and recovery needs
   - Daily activity (steps, calories burned)
   - Exercise preferences and frequency
   - Any dietary restrictions or constraints

5. Specify intensity using heart rate zones and specific sets/reps/rest periods
6. Include proper warm-up and cool-down for each workout day
7. End by asking for feedback about the plan

IMPORTANT: Generate completely personalized plans. Each person's plan should be unique based on their specific health profile, goals, and preferences.

Present everything in clear, well-structured Markdown format.""",
    description="Expert fitness planning agent that creates personalized weekly training plans by directly analyzing comprehensive user health data without intermediate processing tools.",
    tools=[],  # No tools needed - Gemini analyzes data directly
)
